[SDPA-5090] Move Date Time functions to new utils file
