+++ packages/components/Atoms/Global/utils/helpers.js
@@ -130,83 +130,6 @@ const isIPadPro = () => {
   return false
 }
 
-const epochToDateText = (epoch) => {
-  if (!isEpoch(epoch)) return ''
-
-  const monthNames = [
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December"
-  ]
-
-  const date = epochToDate(epoch)
-
-  if (epochIsYesterday(epoch)) return 'Yesterday at ' + formatAMPM(date)
-  else if (epochIsToday(epoch)) return 'Today at ' + formatAMPM(date)
-  else {
-    return String(date.getDate()) +
-      ' ' +
-      String(monthNames[date.getMonth()]) +
-      ' ' +
-      String(date.getFullYear())
-  }
-}
-
-const epochIsYesterday = (epoch) => {
-  const date = epochToDate(epoch)
-
-  // Yesterdays date
-  const yesterday = new Date(new Date())
-  yesterday.setDate(yesterday.getDate() - 1)
-
-  return date.getDate() === yesterday.getDate() &&
-    date.getMonth() === yesterday.getMonth() &&
-    date.getFullYear() === yesterday.getFullYear()
-}
-
-const epochIsToday = (epoch) => {
-  const date = epochToDate(epoch)
-  const nowDate = new Date()
-
-  return date.getDate() === nowDate.getDate() &&
-    date.getMonth() === nowDate.getMonth() &&
-    date.getFullYear() === nowDate.getFullYear()
-}
-
-const epochToDate = (epoch) => {
-  const utcSeconds = epoch
-  const date = new Date(0) // The 0 there is the key, which sets the date to the epoch
-  date.setUTCSeconds(utcSeconds)
-  return date
-}
-
-const isEpoch = (epoch) => {
-  const date = new Date(0) // The 0 there is the key, which sets the date to the epoch
-  date.setUTCSeconds(epoch)
-
-  return !isNaN(date.getTime())
-}
-
-const formatAMPM = (date) => {
-  let hours = date.getHours()
-  let minutes = date.getMinutes()
-  const ampm = hours >= 12 ? 'pm' : 'am'
-  hours = hours % 12
-  hours = hours || 12 // the hour '0' should be '12'
-  minutes = minutes < 10 ? '0' + minutes : minutes
-  const strTime = hours + ':' + minutes + ' ' + ampm
-  return strTime
-}
-
 export {
   capitalize,
   decodeSpecialCharacters,
@@ -217,6 +140,5 @@ export {
   isExternalUrl,
   isIPadPro,
   isRelativeUrl,
-  truncateText,
-  epochToDateText
+  truncateText
 }
